requirements.txt
paho-mqtt
plotly
mysqlclient
mysql-connector-python
pymysql
dash